+++ index.html
@@ -172,7 +172,8 @@
     <hr>
     <p>Created by the <b>Digital Dr3am T3am</b></p>
   </footer>
-  <script src="script.js"></script>
+  <script src="script.js" type="module"></script>
+  <script src="tests.js" type="module"></script>
 </body>
